feat: remove render caching and move props compare logic
